requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.5
sqlalchemy==2.0.23
pyjwt==2.8.0
bcrypt==4.0.1
python-jose==3.3.0
requests==2.31.0
stripe==6.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
